Changed the original player sprite, added an animation player, added the new enemies,. created a hitbox for the enemies so they can attack the player as well
Right now there is a animation player but it doesn't play when the player attacks it just switches to the old picture. then I couldn't animate the walking animatin because its on a different sprite sheet. uhhhh wizard stuns themsevles right now becasue they are spawning a ball on top of themselves and they are getting hit with it instead. enemy3 rock guy has the extra hit box to detect and hurt the player. good luck to all of you ! i have to go to work